Add remaining exceptions
Based on the php library. In case of network failure the error object
does not respond to #code or #message. Let's avoid dealing with that in
the error class and shift the responsability of formatting the error
message to the place where the error is being raised from. Should be
next commit.