Rescue from error when find non existing workflow
The engine response has changed. Before when trying to fetch a non
existing workflow, it would respond with a 200 and a message saying the
requested workflow was not found. Now it returns a 404. While this is
semantincally more correct, our http module treats anything with a HTTP
status code greater than 400 as an error and converts the error to a
Zenaton internal error. This in turn blocks us from rescueing from a
specific error when doing the actual call, so we now check the message
instead. If the message corresponds to a workflow not being found, we
return `nil` as before. If not we re raise the exception.