Fix serialization error with ActiveModel object
When checking if an object is true or false using `==` raised an error
with ActiveModel objects. AM definition of `==`
https://github.com/rails/rails/blob/master/activemodel/lib/active_model/attribute_set.rb#L100
assumes the other object being compared is also an AM object and calls
`#attributes` on it. It is not rare to override the `#==(other)` method
so let's use `is_a?` instead.