Url encode query params (#51)
There was an issue where a workflow which used an email address as an ID
could not be stopped. This was caused by the custom id not being properly
encoded. This changes how params are passed to the client. We now use a
hash and leverage the `URI` module to ensure the hash is correctly
encoded into query params when needed.

The old behaviour is deprecated and users of an old version of the agent will
receive a warning to upgrade the agent.